connected to redux now
